+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   [#303](https://github.com/lambda-fairy/maud/pull/303)
 - Add support for `Option<T>` attributes using the `attr=[value]` syntax.
   [#306](https://github.com/lambda-fairy/maud/pull/306)
+- Update to Rust 2021
+  [#309](https://github.com/lambda-fairy/maud/pull/309)
 
 ## [0.22.3] - 2021-09-27