+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   [#309](https://github.com/lambda-fairy/maud/pull/309)
 - Remove Iron support
   [#289](https://github.com/lambda-fairy/maud/pull/289)
+- Disallow slashes (`/`) in void elements
+  [#315](https://github.com/lambda-fairy/maud/pull/315)
 
 ## [0.22.3] - 2021-09-27