+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.23.0] - 2021-11-10
+
 - Update to support axum 0.2
   [#303](https://github.com/lambda-fairy/maud/pull/303)
 - Add support for `Option<T>` attributes using the `attr=[value]` syntax.
@@ -266,7 +268,8 @@
 - Update to latest syntax extension API
 
 
-[Unreleased]: https://github.com/lambda-fairy/maud/compare/v0.22.3...HEAD
+[Unreleased]: https://github.com/lambda-fairy/maud/compare/v0.23.0...HEAD
+[0.23.0]: https://github.com/lambda-fairy/maud/compare/v0.23.3...v0.23.0
 [0.22.3]: https://github.com/lambda-fairy/maud/compare/v0.22.2...v0.22.3
 [0.22.2]: https://github.com/lambda-fairy/maud/compare/v0.22.1...v0.22.2
 [0.22.1]: https://github.com/lambda-fairy/maud/compare/v0.22.0...v0.22.1